Reorganized files and modified login and sign-up pages
