+++ src/components/Navbars/IndexNavbar.js
@@ -1,5 +1,5 @@
 import React from "react";
-// import { Link } from "react-router-dom";
+
 // reactstrap components
 import {
   Collapse,
