+++ src/views/index-sections/SignUp.js
@@ -28,7 +28,6 @@ function SignUp() {
       <div
         className="section section-signup"
         style={{
-          backgroundImage: "url(" + require("assets/img/bg11.jpg") + ")",
           backgroundSize: "cover",
           backgroundPosition: "top center",
           minHeight: "700px"
